Refactor world and game objects
