+++ lib/car.js
@@ -10,8 +10,8 @@ function Car(x, y, width, height, direction, world) {
 }
 
 Car.prototype.draw = function(context, carRightImage) {
-  context.drawImage(carRightImage, this.x, this.y, this.width, this.height);
-  context.fillStyle = 'transparent';
+  this.world.context.drawImage(carRightImage, this.x, this.y, this.width, this.height);
+  this.world.context.fillStyle = 'transparent';
   return this;
 };
