+++ lib/index.js
@@ -1,4 +1,1 @@
-var Car = require("./car.js");
-var Frog = require("./frog.js");
-var Log = require("./logs.js");
-var World = require("./world.js");
+require("./game.js");
