+++ lib/logs.js
@@ -8,8 +8,8 @@ function Log(x, y, width, height, direction, world) {
 }
 // could we create prototypes for x and y that increment through Car?
 Log.prototype.draw = function(context, logImage) {
-  context.drawImage(logImage, this.x, this.y, this.width, this.height);
-  context.fillStyle = 'transparent';
+  this.world.context.drawImage(logImage, this.x, this.y, this.width, this.height);
+  this.world.context.fillStyle = 'transparent';
   return this;
 };
