+++ test.html
@@ -6,6 +6,6 @@
 </head>
 <body>
   <script src="test.bundle.js"></script>
-  <canvas id="game" width="1000" height="600">test text</canvas>
+  <canvas id="game" width="500" height="700"></canvas>
 </body>
 </html>
