+++ test/index.js
@@ -1,3 +1,4 @@
 require('./car-test.js');
 require('./logs-test.js');
 require('./frog-test.js');
+require('./world-test.js');
